mock-mcp-servers/requirements.txt
# Mock MCP Servers Dependencies
fastapi>=0.109.0
uvicorn>=0.27.0
pyyaml>=6.0.0
pydantic>=2.0.0
httpx>=0.27.0
